Workaroud to run on lxplus
