+++ Instruments/include/DYFileDescriptor.h
@@ -11,17 +11,11 @@ This file is part of https://github.com/hh-italian-group/hh-bbtautau. */
 #include "AnalysisTools/Core/include/PhysicalValue.h"
 #include "h-tautau/Analysis/include/SummaryTuple.h"
 #include "AnalysisTools/Core/include/ConfigReader.h"
+#include "Instruments/include/SampleDescriptor.h"
 
 namespace analysis {
 
 namespace sample_merging{
-ENUM_OSTREAM_OPERATOR()
-ENUM_ISTREAM_OPERATORS()
-enum class FileType { inclusive, exclusive };
-ENUM_NAMES(FileType) = {
-    { FileType::inclusive, "inclusive" },
-    { FileType::exclusive, "exclusive" }
-};
 
 struct DYBinDescriptor {
     std::string name;
