+++ Instruments/include/SampleDescriptor.h
@@ -14,6 +14,15 @@ This file is part of https://github.com/hh-italian-group/hh-bbtautau. */
 namespace analysis {
 
 namespace sample_merging{
+#ifndef __APPLE__
+    ENUM_OSTREAM_OPERATOR()
+    ENUM_ISTREAM_OPERATORS()
+#endif
+enum class FileType { inclusive, exclusive };
+ENUM_NAMES(FileType) = {
+    { FileType::inclusive, "inclusive" },
+    { FileType::exclusive, "exclusive" }
+};
 
 template<typename BinDescriptor, typename GenMap>
 struct SampleDescriptor {
